corrected rename imagesDir.js to viewDirs.js
